requirements.txt
langchain-openai==0.1.0
redis==5.0.1
python-dotenv==1.0.0
colorama==0.4.6
prompt-toolkit==3.0.43
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
